pyproject.toml
[project]
name = "accounting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-mgmt-carbonoptimization>=1.0.0",
    "python-dateutil>=2.9.0.post0",
    "jupyter>=1.1.1",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-compute>=1.19.0",
    "google-cloud-resource-manager>=1.12.0",
    "google-cloud-service-usage>=1.10.0",
    "google-cloud-logging>=3.10.0",
    "boto3>=1.35.0",
]
